requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
altair==5.5.0
altgraph==0.17.4
annotated-types==0.5.0
anyio==3.7.1
apify_client==1.8.1
apify_shared==1.2.1
arabic-reshaper==3.0.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bottle==0.13.2
bs4==0.0.2
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.6
clr_loader==0.2.7.post0
colorama==0.4.6
comtypes==1.4.9
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
Cython==3.0.11
daphne==4.0.0
distlib==0.3.9
distro==1.9.0
dronekit==2.9.2
dronekit-sitl==3.3.0
exceptiongroup==1.2.2
fastapi==0.101.0
ffmpeg-python==0.2.0
filelock==3.16.1
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
future==0.18.3
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.0
google-auth==2.37.0
google-cloud-core==2.4.1
google-cloud-datastore==2.20.2
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
iniconfig==2.0.0
iso8601==2.1.0
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==4.9.3
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
monotonic==1.6
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
narwhals==1.21.1
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==1.26.4
openai==0.28.0
openai-whisper==20231117
opencv-python==4.10.0.84
outcome==1.3.0.post0
packaging==21.3
pandas==2.2.3
pefile==2024.8.26
pexpect==4.9.0
pillow==11.1.0
pkginfo==1.9.6
platformdirs==2.6.2
playwright==1.49.1
plotly==5.24.1
pluggy==1.5.0
pocketsphinx==5.0.4
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.2
proxy-tools==0.1.0
psutil==6.1.1
ptyprocess==0.7.0
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycountry==24.6.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==1.10.20
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
pygame==2.6.1
Pygments==2.19.1
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2022.14
pymavlink==2.4.37
pynmeagps==1.0.46
pyOpenSSL==24.3.0
pyparsing==3.2.1
pyrtcm==1.1.4
pyserial==3.5
pyshorteners==1.0.1
PySocks==1.7.1
pystray==0.19.5
pytest==8.3.4
python-barcode==0.15.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.3.4
python-pptx==1.0.2
python-socketio==5.7.2
pythonnet==3.0.5
pytz==2024.2
pyubx2==1.2.50
pywebview==4.3.3
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
safetensors==0.5.2
seaborn==0.13.2
service-identity==24.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.0
sortedcontainers==2.4.0
sounddevice==0.4.6
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.21
sse-starlette==2.1.3
starlette==0.27.0
streamlink==7.1.2
streamlit==1.41.1
sympy==1.13.1
tenacity==9.0.0
textblob==0.18.0.post0
tiktoken==0.8.0
tokenizers==0.21.0
toml==0.10.2
tomli==2.2.1
torch==2.2.1+cu118
torchaudio==2.2.1+cu118
torchvision==0.20.1+cu118
tornado==6.4.2
tqdm==4.67.1
transformers==4.48.0
trio==0.28.0
trio-websocket==0.11.1
ttkthemes==3.2.2
Twisted==24.11.0
twitchio==2.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.20.0
virtualenv==20.16.7
watchdog==6.0.0
websocket-client==1.8.0
websockets==10.4
Werkzeug==3.1.3
wordcloud==1.9.4
wsproto==1.2.0
XlsxWriter==3.2.2
yarl==1.18.3
zope.interface==7.2
